requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.2

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Database
sqlalchemy>=2.0.0

# XML parsing (standard library, but listed for clarity)
# xml.etree.ElementTree is built-in

# Date/time handling
pytz>=2023.3

# Optional: Jupyter for notebooks
jupyter>=1.0.0
ipykernel>=6.25.0

